+++ src/embedProcessor.ts
