+++ src/types/inline.d.ts
@@ -1,0 +1,4 @@
+declare module 'inline:*' {
+    const content: string;
+    export default content;
+}
